Include graphic in chat-message output
