+++ system/templates/rolls/chat-message.hbs
@@ -1,3 +1,6 @@
-<div class="ironsworn-roll" data-ironswornroll="{{ironswornroll}}">
+<div class="ironsworn-roll flexcol" data-ironswornroll="{{ironswornroll}}">
+  <div class="dice-result dice-tooltip dice-roll ironsworn-roll">
+    {{{graphic}}}
+  </div>
   <pre><code>{{json ironswornroll}}</code></pre>
 </div>
