+++ system/templates/rolls/roll-graphic.hbs
@@ -1,4 +1,4 @@
-<div class="roll dice-roll ironsworn-roll">
+<div class="dice-roll ironsworn-roll">
 
   <div class="dice-formula">
     <p>
